Created board-specific hal files (#139)
# Purpose
Since the RM46 is configured differently between the different boards
(launchpad, rev1, and rev2), it's tough to configure the HAL (using
halcogen) to work with all boards. It's only gonna get more difficult as
we get new boards.

I know we talked about this in a CDH meeting and decided to create
separate hal files for each board. This PR implements that. But idk if
this'll scale well (there's a ton of files that are common between the
board-specific hal directories).

Alternatively, we can write our own code to configure the rm46 on boot
(instead of relying on halcogen-generated code). This would be more
difficult tho

# New Changes
- Created separate hal directories.
- Updated build scripts
- Added RTC alarm fix

# Testing
- CI

# Outstanding Changes
- This will break the old build setup that doesn't use CMake. We should
just remove those files now