Navtaj/comms protocol reorder (#155)
# Purpose
Explain the purpose of the PR here, including references to any existing
Notion tasks.
- With the current order of comms protocols, we need to ignore the FCS
field of ax25 which leads to undetectable errors in the ax25 header
fields. This can be an issue if the error is in the control field, so
this new order performs FEC on the info field before ax25Recv so that we
can use the checksum to detect any errors with the ax25 headers
- Added unit tests for ax25
- Added a mock logging file to define a separate logLog function for
testing that does not interact with HAL
# New Changes
- Explain new changes

# Testing
- Explain tests that you ran to verify code functionality.
- CI
- Unit Tests passed
# Outstanding Changes
- If there are non-critical changes (i.e. additional features) that can
be made to this feature in the future, indicate them here.