Directory revamp (#149)
# Purpose
Change to the directory structure outlined here:
https://www.notion.so/uworbital/FW-Next-Comp-Planning-f04acfffbc6c49fc86d8486fd48e70ab?pvs=4#026f8eb4d991459a8f8ff73fb6876d21
**Note: The Notion page might be out-of-date**

**Goal:** 
- organize the repo so we can put all projects (gs, OBC, etc) in one
repo.
- update the build system to handle tests, OBC fw, and gs code. Our
current build system doesn't use CMake to build the gs.
- Decouple the obc_gs_interface from the OBC fw
- Split things up into more logical sections. I wanted to make sure we
didn't have a situation like last comp where `cdh/` had different things
(command manager, telemetry manager, timekeeper, supervisor) and we
dumped a bunch of stuff into `common/` (some of which can be used by the
ground station like packing utils and others are tied to the OBC fw like
obc_time)

# New Changes
- Changed directory structure; I tried to minimize the number of folders
at the top-level of the directory.
- Switched from include guards to `#pragma once` for a few reasons:
  1. It makes refactoring simpler
2. It's less prone to bugs since you don't have to choose a unique
include guard (that can sometimes be a lot of characters)
  3. It's less code to write
- Moved interface-related comms files to `obc_gs_interface` and added
the `obc_gs_` prefix to each filename
- Added project-specific error codes. `obc_error_code_t` for OBC
firmware, `obc_gs_error_code_t` for interface code, and
`gs_error_code_t` for ground station code. Because of this, I've removed
the use of `RETURN_IF_ERROR_CODE` from the interface for now. In a
future PR, we'll add a logging module that'll work across the obc, gs,
and interface.
- Created standalone obc_gs_interface lib that can be linked into any
executable (obc, gs, or test)
- Removed any dependencies on the main firmware from the
obc_gs_interface module (this allows us to compile it as a static lib)

Other stuff:
- Added telemetry unpack functions
- Added some unit tests (for telemetry pack and obc_time_utils)
- Moved ISR callbacks to `drivers/` to keep the HAL independent from the
rest of our code. This should, in theory, allow us to compile the HAL as
a static lib. I also added back the rtc alarm callback which was deleted
accidentally in a past PR.

# Testing
- CI
- Tested fw on launchpad

# Outstanding Changes
- The CC1120 driver shouldn't be dependent on code from `services/`
- Make a static lib for the hal and then fix the builds for all the test
apps. It'd be nice if we could just link the HAL static lib when
building each test app.
- Create a logging module (with RETURN_IF_ERROR_CODE macros) that'll
work with the different error codes we have
- Write CMake and unit test documentation
- Combine obc_state into supervisor and rethink what telemetry we send
down.