Sun Position Ephemeris Ground Station (#183)
# Purpose
Create the ground station side of the sun position ephemeris algorithms
required by GNC
Task Board:
https://www.notion.so/uworbital/Sun-Position-Ephemeris-Algorithms-Ground-Station-7b1692a31c544785be143bfca1083c3d

# New Changes
- Added script `ephemeris.py` to create a binary file with the data
points required by GNC for algorithms on the OBC. The binary file will
be read and data sent to the OBC (comms handle this).
- Added parser script `ephemerisparser.py` to parse the output files
created by the ephemeris.py, this is a test harness and will be the
basis for the parser on the OBC
- Added test files to validate the above scripts located in the
`gs/sun/test` directory
- Added `README.md` to explain the usage of `ephemeris.py`
- Added the requirements for the scripts and testing to the
`requirements.txt` file
- Added directories for Python and Pytest caching to the `.gitignore` 

# Testing
- Tests run in `gs/sun/test` using Pytest, all tests passed (113 unit
tests)
- Additional tests were performed in the command line for `ephemeris.py`
to check the 'exclude' -e flag

## The functions that weren't unit tested
- `ephemeris.py::print_debug_header` as its only purpose is to make it
easier to read debug messages by printing a debug table header, which
doesn't affect the main script. Also, it would be impractical to test as
it logs multiple times lines
- `ephemeris.py::validate_response` as it requires a Response object
that can't be simulated without making bad requests on purpose.

# Outstanding Changes
- Add a CLI to the `ephemerisparser.py` to parse the output files of
`ephemeris.py`. This would be nice to add but not necessary right now as
you can use `-p 1` or `-p 2` when running `ephemeris.py` to validate the
written data
- Add an internal Python package for code that isn't specific to a task
such as `ephemeris.py::is_float` or
`test_ephemeris.py::suppress_logging`

# New after review
- Fixed typos
- Added testing for `test_ephemeris.py::suppress_logging`