Download correct v12.3 toolchain if Apple Silicon (#197)
# Purpose
Currently, OBC-firmware cannot be built on Apple Silicon chips. This is
because support for Apple Silicon was added to the ARM embedded
toolchain from v12.2+. OBC-firmware is currently on v11.2, and does not
download the correct toolchain for Apple Silicon devices.

# New Changes
Change `download_arm_toolchain.cmake` to determine CPU architecture and
download the correct toolchain for Apple devices.

# Testing
OBC-Firmware was built successfully on an M2 MacBook.

# Outstanding Changes
N/a.