Adcs/pratyush/h bridge (#182)
# Purpose
Explain the purpose of the PR here, including references to any existing
Notion tasks.

https://www.notion.so/uworbital/6375c5c5dcf1485e8a6c47f18d419ae5?v=72baf90d01f643b09d77418bd3f53b14&p=6c975a20ccca4e29b449891976e27d30&pm=s

A driver for the H-Bridge motor. 
# New Changes
Added functions to drive the motor forward, reverse, brake and set it to
idle mode using the HET PWM hal library.

# Testing
- Explain tests that you ran to verify code functionality.
- Any functions that can be unit-tested should include a unit test in
the PR. Otherwise, explain why it cannot be unit-tested.
I do have to test it using the board, would likely need to go into the
bay to do that sometime this week. There is little to unit test since
the driver just modifies PWM signals.

# Outstanding Changes
- If there are non-critical changes (i.e. additional features) that can
be made to this feature in the future, indicate them here.
SInce I have a Mac, someone would need to configure the PWM for
HalCoGen. Will make a ticket for that.