Yarik/generalize persist (#240)
# Purpose
Generalize the persistent storage. There will be no need to write custom
getters and setters for each persistent section.

# New Changes
- Created 2 getters (`getPersistentDataByIndex` for sections of length
greater than or equal 1 and `getPersistentData` which can be used for
the sections of length 1)
- Created 2 setters (`setPersistentDataByIndex` for sections of length
greater than or equal 1 and `setPersistentData` which can be used for
the sections of length 1)
- Removed previous getters/setters
- Updated all occurrences of previous getters/setters 
- Added persistent config for storing the necessary data for
generalization
- Updated instructions for the persistent
- Changed `framWrite` to take const buffer
- Added mock crc and fram for SIL testing
- Updated the mock logging as it wasn't building the tests

# Testing
- SIL tests in `test_obc_persistent.cpp`
- Ran on rev1 board using the
`obc/examples/test_app_fram_persist/main.c` and got the expected output

# Outstanding Changes
- Fram dump command
- Decouple obc_persistent from alarm and command managers (i.e.
non-persistent elements)

# Comments/Questions
- The `mock_crc.c` file is just a copy and paste of the mock from the
crc.c file with the copyright details and red function calls removed.
Maybe there's a better way of doing it where we would remove the red
function calls from the original crc.c file when building the tests?
- The buffLen param for the getters and setters is only used to check
that the buffer is long enough. All other operations (read, write, crc
calculations are done on the dataSize). Does it make sense to keep it?
- Should the buffer be changed to a void* as its kinda a mess to call
the getters/setters without them and it seems that the calls to them
will always just be a pointer to some persistent data struct?
- For the fram dump command, will we want to take the data by section
(time, alarm manager etc.) or just everything without caring which
section the data comes from? If we are doing option 1 then, we will need
to add a macro for the maximum data size and it will need to be checked
by comparing if its greater of equal to the largest data size in the
configuration. This is the way I'm thinking of doing it: an
initPersistent will need to be added that asserts if the macro is large
enough to store the sections, changing a static variable to indicate
that it was tested. If it wasn't tested then all persistent calls will
return invalid state error