VN100_Driver (#237)
# Purpose
Create a driver capable for reading from and configuring the vn-100 with
abstracted functions for ASCII commands.

# New Changes
- Added two test apps for serial ASCII and serial binary outputs.
- Added ascii and binary output configurations and parsing of binary
output packets to obtain float values for (yaw, pitch, roll), Angular
rates, acceleration, magnetometer readings, temperature and pressure.

# Testing
- Ran both of the example codes that I created to test ASCII and binary
outputs as well as the parsing of the packets.
- Tested ASCII configuration commands using RS-232 serial interface with
laptop
- Created unit tests with mock packets to test the parsing and unpacking
of VN100 packets.

# Outstanding Changes
- Re-implement checksum error checking for packets
- Implement a function to handle serial inputs from user to configure
the VN-100. (e.g I type "$VNRRG,06*XX" into the serial terminal to send
to the RM46, which should send to the VN-100 so that I can read from
register 6 of the VN-100)
- Mainly want to implement this because it allows us to make on the fly
configurations as necessary without having to build new wrappers.