Command callback (#264)
# Purpose
Command callback PR for updating the ground station with a response for
commands sent alongside telemetry.

# New Changes
Implemented pack/unpack interface for the telemetry commands. Also
changed the macros of command IDs to enum type.

# Testing
Developed a range of cpp tests. 

# Outstanding Changes
- If there are non-critical changes (i.e. additional features) that can
be made to this feature in the future, indicate them here.