Upgrade to Python 3.11 (#316)
# Purpose
When Python 3.10 was chosen for GS, it was expected that work on the
backend would start a lot sooner i.e. S24 when version 3.11 was still in
it's bug fix stage. Since, the start of work on the backend was delayed
until W25, Python 3.11 is now in stable/security releases. As such it is
now considered stable for GS development.

This PR upgrades the main repo to use 3.11. This version adds new
features and improved performance with very little that must be done by
GS developers to migrate (See below for Migration instructions). Also,
the end of life has been extend by 1 year from Oct 2026 (Python 3.10) to
Oct 2027 (Python 3.11).

# Migration Instructions
1. Delete your existing python virtual environment directory used for
this repo. Typically the directory is called `.venv` or `venv`
2. Follow the setup instructions for the backend in the README
- Step 2 of the setup instructions for MacOS and Linux users
- Step 3B of the setup instructions for WSL Windows

# New Changes
- Update README with instructions on how to install Python 3.11
- Update python configuration files to use version 3.11
- Update Github actions to use Python 3.11 where applicable
- Minor README improvements

# Testing
- CI/CD
- Backend starts successfully

![image](https://github.com/user-attachments/assets/4bcb60fc-ed30-443d-8e07-a1d10afc750b)

# Outstanding Changes
- Upgrade Python when 3.11 approaches end of life (good to upgrade
around W27 or S27 at the latest) or sooner if desirable

# Additional Notes
- Upgraded GS onboarding Python 3.11 version with this PR:
https://github.com/UWOrbital/GS-Onboarding/pull/12