Fix issue templates directory (#329)
# Purpose
The templates weren't in the correct directory.

# New Changes
- Fix issue templates directory name

# Testing
Validated that it works by trying to create an issue from the template
inside of GS projects and replacing `main` inside of
`https://github.com/UWOrbital/OBC-firmware/tree/main/.github/ISSUE_TEMPLATE`
with `yarik/fix-issue-templates-dir` to get access to the issue
templates as desired.

# Outstanding Changes
Same as #326