Update issue templates (#331)
# Purpose
Hopefully finally fix our broken issue templates.

Create issues via the issue template builder. I'm honestly getting
really tired of this. It's ridiculous that you can't test this feature
on a separate branch, and the documentation isn't great.

# Testing
The only way to test this was to change the default branch from `main`
to this branch (which is really dumb) and then go into the project.
Luckily, the templates show up and they are what I made them to be. I've
switched the default branch back to `main`.


![IMG_20250309_192958_144.jpg](https://github.com/user-attachments/assets/4bf89742-a757-4738-b1d5-59c0372dfea5)

# Outstanding changes
Hopefully none