Modify command policy handling (#374)
# Purpose
Basically just makes a minor change to how command policies are
assigned/work.
Task:
https://uworbital.notion.site/Modify-command-policy-handling-aa594df7180d4bc2bb22db11193551c5

# New Changes
- When creating commands you just set the command policy to the most
restrictive policy you want. Rn all of our commands are available for
PROD anyways. But PROD is now a subset of RND, so all PROD commands will
work in RND.

# Testing
- Built and didn't break anything :)
- Confirmed the OBC_ACTIVE_POLICY build flag works and is accessible in
the code
- I actually don't know how to formally test this because its part of a
manager/task, so it's not like you can do unit tests. But the changes
are pretty minor so...

# Outstanding Changes
- N/A