Implement scheduler config generation (#375)
# Purpose
Notion task:
https://github.com/UWOrbital/OBC-firmware/blob/main/obc/app/modules/digital_watchdog_mgr/digital_watchdog_mgr.c#L38-L102

Right now, creating a new task involves adding a bunch of different
macros to the obc_scheduler_config file as well as calling the task
create function in state_mgr and adding a timeout for the task in
watchdog_mgr.
We generate new scheduler config c and header files with a python
script, which parses a toml file.

# Warning: Breaking Changes
All firmware developers will need Python installed and their virtual
environment activated to build the obc or the examples

# New Changes
- Added skeleton files to generate config from a template
- Python script parses scheduler_config.toml to generate config from
skeleton files (we use toml as it supports comments)
- Macros are redundant, so we use their values directly
- Added warnings to not modify generated code
- Use jinja2 for template support
- Imported jinja2, toml, and types-toml
- Header file support

# Testing
- Built successfully
- Flashed to board
- Scheduler config files should be functionally identical
- Proper field validation

![image](https://github.com/user-attachments/assets/b49c120e-8135-4ee0-8b31-1a42f170c4ad)
- Unit tests not applicable

# Outstanding Changes
- Also generate task timeouts in digital_watchdog_mgr.c
- Pull macros from config as well instead of having them as comments,
then we can add additional validation

---------

Co-authored-by: Yarik Popov <89220488+Yarik-Popov@users.noreply.github.com>