Add binary neutrality check (#434)
# Purpose
We should have a workflow that checks if a PR changed the firmware
executable. It'd be useful when making changes that shouldn't affect the
firmware (but accidentally do).
We create workflows for main and PRs that calculate and display the CRC
of the .bin file; we would be manually comparing if the CRC changed. The
task originally said to check .out files.
Closes https://github.com/UWOrbital/OBC-firmware/issues/429

# New Changes
- We now upload artifact containing the binary after build
- Added "crc" job to OBC Build workflow to display CRC checksum of built
binary, downloading artifact uploaded from before
- Removed build timestamp to ensure deterministic builds

# Testing
- Tested in a fork of the repository. Also, this PR itself acts as a
test

# Outstanding Changes
- N/A