Fix FRAM spi format that was breaking SDC IO (#505)
# Purpose
This is to fix the below SD card IO error. It was due to the FRAM driver
using `SPI_FMT_1` which was likely causing some unintended bus activity
that was affecting the SD card IO (need to test with a logic analyzer to
be certain though). SD card IO worked fine right after booting, but as
soon as the timekeeper task ran (which was setting persistent data in
the FRAM), the SD card IO failed. Using the same SPI_FMT on both tasks
fixed it (both FRAM and SD card reads and writes worked)

```
23-01-01_00-00-00 ERROR -> obc/app/sys/fs_wrapper/obc_reliance_fs.c:38 - 1005
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_fs_utils.c:17 - 710
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_manager.c:68 - 710
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_fs_utils.c:33 - 701
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_fs_utils.c:94 - 701
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_manager.c:71 - 701
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_fs_utils.c:123 - 2
23-01-01_00-00-00 ERROR -> obc/app/modules/telemetry_mgr/telemetry_manager.c:77 - 2
```

# New Changes
- Modify board_config to use SPI_FMT_0 for FRAM in REV1

# Testing
- Created test task to test FRAM and SD card read and write. See [this
branch](https://github.com/UWOrbital/OBC-firmware/compare/main...thenuja/bugfix/sdc_mkdir)
- Output from test task:
```
############################################ Running FRAM + SDC integration test!
======     Testing FRAM read/write   ======
Wrote 69420 to FRAM.
Read 69420 from FRAM.
PASSED FRAM TEST!
====== Finished FRAM write/read test ======
======     Testing SDC read/write   ======
Wrote follwoing string to /sdcTestDir/testfile.txt and closed the file:
---
This is a test string that is written to SDC!
---
Opened /sdcTestDir/testfile.txt and read 56 bytes, expected 56 bytes.
Read string:
---
This is a test string that is written to SDC!
---
PASSED SDC TEST!
====== Finished SDC write/read test  ======
```

# Outstanding Changes
- NA