Reduced BL size (#504)
# Purpose
This PR closes #427 
This PR reduces the size of the bootloader down so that it does not
exceed the `0x20000` address (131072 bytes in size)

> [!NOTE]
> While looking at the cmake files, it seems that `FreeRTOS` is just
included in the HAL library used in the bootloader but the sources are
not included. I assume that means it get's optimized out. I didn't want
to remove the includes since it seems that the HAL uses it and I'd have
to tinker with the autogenerated HAL files.

# New Changes
- Removed `BL_KERNEL`
- Added `-Os` to cmake to optimize for size

# Testing
Explain tests that you ran to verify code functionality.
- [x] I have unit-tested this PR. Otherwise, explain why it cannot be
unit-tested.
- Instead of unit testing, I have tested the new bootloader script and
verified that it works with the new bootloader layout
- I have verified via the Uniflash memory reader that the bootloader
program does not exceed the `0x00020000` memory address.
- [x] I have tested this PR on a board if the code will run on a board
(Only required for firmware developers).
- [ ] I have tested this PR by running the ARO website (Only required if
the code will impact the ARO website).
- [ ] I have tested this PR by running the MCC website (Only required if
the code will impact the MCC website).
- [x] I have included screenshots of the tests performed below.
<img width="1893" height="1002" alt="image"
src="https://github.com/user-attachments/assets/320d3321-ffdd-4c47-81f5-5a78266ae203"
/>
Memory verification via Uniflash

# Outstanding Changes
 - None as of creation

---------

Co-authored-by: Adityya Kaushal <52387864+Adityya-K@users.noreply.github.com>
Co-authored-by: Adityya Kaushal <adityyakaushal@gmail.com>