Add environment files to aro and MCC frontend (#546)
# Purpose
Closes #343.
Explain the purpose of the PR here if it doesn't match the linked issue.
Be sure to add a comment in the linked issue explaining the changes.

# New Changes
Explain new changes below in short bullet points.
- Created .env files for aro and MCC frontend

# Testing
Explain tests that you ran to verify code functionality.
- [ ] I have unit-tested this PR. Otherwise, explain why it cannot be
unit-tested.
- [ ] I have tested this PR on a board if the code will run on a board
(Only required for firmware developers).
- [ ] I have tested this PR by running the ARO website (Only required if
the code will impact the ARO website).
- [ ] I have tested this PR by running the MCC website (Only required if
the code will impact the MCC website).
- [ ] I have included screenshots of the tests performed below.

# Outstanding Changes
If there are non-critical changes (i.e. additional features) that can be
made to this feature in the future, indicate them here.
- Values for each .env file must be filled
- More urls and keys to be added
- Divide .env into .env.development/.env.production?

Co-authored-by: Adityya Kaushal <52387864+Adityya-K@users.noreply.github.com>