Update README in favour of docs site (#560)
# Purpose
Heavily cut down on README content as part of the transition to the docs
website. Some guides were outdated or incorrect, while others were
formatted inconsistently. These changes are present on the docs site.

# New Changes
- Added the UW Orbital Docs section to the top of the document to
highlight that it contains essential sub-team information
- Deleted large sections of the README, including the setup guides,
build guides, and style guides, which are now found on the documentation
website
- Made some minor formatting changes to some remaining sections
- Removed "Back to top" hyperlinks since the document is short enough
now
- Updated the name of the team from firmware to software
- Added Notion section and links to useful pages