Integrate gs building code for MacOS systems (#552)
# Purpose
Previously, no instructions were provided on how to build the ground
station program for MacOS systems. This PR adds steps on how to fix the
errors that are encountered while building it.

# New Changes
- Instructions on building gs for MacOS systems were added to the readme

# Testing
Explain tests that you ran to verify code functionality.
- [x] I have unit-tested this PR. Otherwise, explain why it cannot be
unit-tested.
- [ ] I have tested this PR on a board if the code will run on a board
(Only required for firmware developers).
- [ ] I have tested this PR by running the ARO website (Only required if
the code will impact the ARO website).
- [ ] I have tested this PR by running the MCC website (Only required if
the code will impact the MCC website).
- [ ] I have included screenshots of the tests performed below.

# Outstanding Changes
If there are non-critical changes (i.e. additional features) that can be
made to this feature in the future, indicate them here.
- Currently, the instructions allow for successfully launching the cli.
More thorough testing may be required for full functionality