refactored wrappers to clarify id type (#598)
# Purpose
Closes #ISSUE_NUMBER.
Explain the purpose of the PR here if it doesn't match the linked issue.
Be sure to add a comment in the linked issue explaining the changes.

# New Changes
Explain new changes below in short bullet points.

- Previously, the object id passed in for delete and retrieve for all
concrete wrappers either could be int or UUID.
- However, the only tables that use int as primary key are the main
table.
- I set the abstract wrapper to take an unspecified type on the delete
and retrieve methods.
- I then specified the param type for those methods in
MainCommandWrapper and MainTelemetryWrapper to only take int.

# Testing
Explain tests that you ran to verify code functionality.
- [ ] I have unit-tested this PR. Otherwise, explain why it cannot be
unit-tested.
- [ ] I have tested this PR on a board if the code will run on a board
(Only required for firmware developers).
- [ ] I have tested this PR by running the ARO website (Only required if
the code will impact the ARO website).
- [ ] I have tested this PR by running the MCC website (Only required if
the code will impact the MCC website).
- [ ] I have included screenshots of the tests performed below.

# Outstanding Changes
If there are non-critical changes (i.e. additional features) that can be
made to this feature in the future, indicate them here.