Retrieve floating commands and add abstract wrappers to conftest (#600)
# Purpose
Closes #574
Created a method under CommandsWrapper which retrieves all current
"floating" commands. A floating command here is defined as any command
whose id does not match up with any command_id in the PacketCommands
table.

# New Changes
retrieve_floating_commands() -> list (Commands)
Extended the conftest to accept the abstracted wrappers
Wrote pytests for it 

# Testing
Pytests include 2 cases: One where PacketCommands is empty and one where
PacketCommands is nonempty.
For the empty one, we expect that after calling
retrieve_floating_commands(), we get the all the commands currently in
the Command table. For the non empty PacketCommands, we expect the
complement of PacketCommands and Commands where PacketCommands is the
universal set :).